neovim(lazy): configure treesitter as a dev module
Configure treesitter as a local dev module as if we were working on changes in a fork of it. This way, lazy will not try and update it, but dependents are still satisfied. We configure lazy to look for dev modules in a local nvim share directory, and configure a symlink to treesitter in that directory via home-manager.